Working process creation and flow, childs able to communicate with parent only once, need to figure out how to flush the pipe
Parent need to retrieve each child info, and print after it finish, order of execution working

Missing, track wait, execute and io time. Sys calls to access that infos about the process and pass to parent.

Parent receive type of process and is able to add the number of processes and total value to the correspondent struct