Sanity working! Need to be tested to make sure
New logic to track ticks, and preempt every 5 ticks. Last logic was wrong and was slowing down SO, now it is correct, tested and validated.

Added new syscall getZombieChildsInfo, that search for a zombie child and retrieve its info. So the parent(sanity) can collect data and return procs averages